(refactor) settings page
